Migrate RFC5 work to v06
