--- src/ome_zarr_models/_rfc5_transforms/axes.py
@@ -1,3 +1,0 @@
-from ome_zarr_models.v05.axes import Axes, Axis, AxisType
-
-__all__ = ["Axes", "Axis", "AxisType"]
