--- src/ome_zarr_models/_rfc5_transforms/base.py
@@ -1,3 +1,0 @@
-from ome_zarr_models.v05.base import BaseGroupv05, BaseOMEAttrs
-
-__all__ = ["BaseGroupv05", "BaseOMEAttrs"]
