--- src/ome_zarr_models/_rfc5_transforms/hcs.py
@@ -1,3 +1,0 @@
-from ome_zarr_models.v05.hcs import HCS
-
-__all__ = ["HCS"]
