--- src/ome_zarr_models/_rfc5_transforms/image_label.py
@@ -1,3 +1,0 @@
-from ome_zarr_models.v05.image_label import ImageLabel
-
-__all__ = ["ImageLabel"]
