--- src/ome_zarr_models/_rfc5_transforms/labels.py
@@ -1,3 +1,0 @@
-from ome_zarr_models.v05.labels import Labels
-
-__all__ = ["Labels"]
