--- src/ome_zarr_models/_rfc5_transforms/plate.py
@@ -1,3 +1,0 @@
-from ome_zarr_models.v05.plate import Plate
-
-__all__ = ["Plate"]
