--- src/ome_zarr_models/_rfc5_transforms/well.py
@@ -1,3 +1,0 @@
-from ome_zarr_models.v05.well import Well
-
-__all__ = ["Well"]
