+++ tests/_v06/test_image_label.py
@@ -1,10 +1,12 @@
 from zarr.abc.store import Store
 
 from ome_zarr_models._v06.axes import Axis
-from ome_zarr_models._v06.coordinate_transformations import VectorScale
+from ome_zarr_models._v06.coordinate_transformations import CoordinateSystem, Scale
 from ome_zarr_models._v06.image_label import ImageLabel, ImageLabelAttrs
 from ome_zarr_models._v06.image_label_types import Color, Label, Source
 from ome_zarr_models._v06.multiscales import Dataset, Multiscale
+from ome_zarr_models.common.coordinate_transformations import VectorScale
+from ome_zarr_models.common.image_label_types import Property
 from tests._v06.conftest import json_to_zarr_group
 
 
@@ -30,6 +32,7 @@ def test_image_label(store: Store) -> None:
     )
     ome_group = ImageLabel.from_zarr(zarr_group)
     assert ome_group.attributes.ome == ImageLabelAttrs(
+        version="0.6",
         image_label=Label(
             colors=(
                 Color(label_value=0, rgba=(0, 0, 128, 128)),
@@ -53,42 +56,76 @@ def test_image_label(store: Store) -> None:
         ),
         multiscales=[
             Multiscale(
-                axes=[
-                    Axis(name="t", type="time", unit="millisecond"),
-                    Axis(name="c", type="channel", unit=None),
-                    Axis(name="z", type="space", unit="micrometer"),
-                    Axis(name="y", type="space", unit="micrometer"),
-                    Axis(name="x", type="space", unit="micrometer"),
-                ],
+                coordinateSystems=(
+                    CoordinateSystem(
+                        name="coord_sys0",
+                        axes=[
+                            Axis(name="t", type="time", unit="millisecond"),
+                            Axis(name="c", type="channel", unit=None),
+                            Axis(name="z", type="space", unit="micrometer"),
+                            Axis(name="y", type="space", unit="micrometer"),
+                            Axis(name="x", type="space", unit="micrometer"),
+                        ],
+                    ),
+                    CoordinateSystem(
+                        name="coord_sys1",
+                        axes=[
+                            Axis(name="t", type="time", unit="millisecond"),
+                            Axis(name="c", type="channel", unit=None),
+                            Axis(name="z", type="space", unit="micrometer"),
+                            Axis(name="y", type="space", unit="micrometer"),
+                            Axis(name="x", type="space", unit="micrometer"),
+                        ],
+                    ),
+                ),
                 datasets=(
                     Dataset(
                         path="0",
-                        coordinateTransformations=(
-                            VectorScale(type="scale", scale=[1.0, 1.0, 0.5, 0.5, 0.5]),
-                        ),
+                        coordinateTransformations=[
+                            Scale(
+                                type="scale",
+                                input="/0",
+                                output="coord_sys0",
+                                name=None,
+                                scale=[1.0, 1.0, 0.5, 0.5, 0.5],
+                            )
+                        ],
                     ),
                     Dataset(
                         path="1",
-                        coordinateTransformations=(
-                            VectorScale(type="scale", scale=[1.0, 1.0, 1.0, 1.0, 1.0]),
-                        ),
+                        coordinateTransformations=[
+                            Scale(
+                                type="scale",
+                                input="/1",
+                                output="coord_sys0",
+                                name=None,
+                                scale=[1.0, 1.0, 1.0, 1.0, 1.0],
+                            )
+                        ],
                     ),
                     Dataset(
                         path="2",
-                        coordinateTransformations=(
-                            VectorScale(type="scale", scale=[1.0, 1.0, 2.0, 2.0, 2.0]),
-                        ),
+                        coordinateTransformations=[
+                            Scale(
+                                type="scale",
+                                input="/2",
+                                output="coord_sys0",
+                                name=None,
+                                scale=[1.0, 1.0, 2.0, 2.0, 2.0],
+                            )
+                        ],
                     ),
                 ),
                 coordinateTransformations=(
-                    VectorScale(type="scale", scale=[0.1, 1.0, 1.0, 1.0, 1.0]),
+                    VectorScale(
+                        type="scale",
+                        scale=[0.1, 1.0, 1.0, 1.0, 1.0],
+                        input="coord_sys0",
+                        output="coord_sys1",
+                    ),
                 ),
                 metadata={
-                    "description": (
-                        "the fields in metadata depend on the downscaling "
-                        "implementation. Here, the parameters passed to the skimage "
-                        "function are given"
-                    ),
+                    "description": "the fields in metadata depend on the downscaling implementation. Here, the parameters passed to the skimage function are given",
                     "method": "skimage.transform.pyramid_gaussian",
                     "version": "0.16.1",
                     "args": "[true]",
@@ -98,5 +135,4 @@ def test_image_label(store: Store) -> None:
                 type="gaussian",
             )
         ],
-        version="0.6",
     )
