+++ tests/data/examples/v06/image_label_example.json
@@ -32,12 +32,28 @@
     "multiscales": [
       {
         "name": "example",
-        "axes": [
-          { "name": "t", "type": "time", "unit": "millisecond" },
-          { "name": "c", "type": "channel" },
-          { "name": "z", "type": "space", "unit": "micrometer" },
-          { "name": "y", "type": "space", "unit": "micrometer" },
-          { "name": "x", "type": "space", "unit": "micrometer" }
+        "type": "gaussian",
+        "coordinateSystems": [
+          {
+            "name": "coord_sys0",
+            "axes": [
+              { "name": "t", "type": "time", "unit": "millisecond" },
+              { "name": "c", "type": "channel" },
+              { "name": "z", "type": "space", "unit": "micrometer" },
+              { "name": "y", "type": "space", "unit": "micrometer" },
+              { "name": "x", "type": "space", "unit": "micrometer" }
+            ]
+          },
+          {
+            "name": "coord_sys1",
+            "axes": [
+              { "name": "t", "type": "time", "unit": "millisecond" },
+              { "name": "c", "type": "channel" },
+              { "name": "z", "type": "space", "unit": "micrometer" },
+              { "name": "y", "type": "space", "unit": "micrometer" },
+              { "name": "x", "type": "space", "unit": "micrometer" }
+            ]
+          }
         ],
         "datasets": [
           {
@@ -45,7 +61,9 @@
             "coordinateTransformations": [
               {
                 "type": "scale",
-                "scale": [1.0, 1.0, 0.5, 0.5, 0.5]
+                "scale": [1.0, 1.0, 0.5, 0.5, 0.5],
+                "input": "/0",
+                "output": "coord_sys0"
               }
             ]
           },
@@ -54,7 +72,9 @@
             "coordinateTransformations": [
               {
                 "type": "scale",
-                "scale": [1.0, 1.0, 1.0, 1.0, 1.0]
+                "scale": [1.0, 1.0, 1.0, 1.0, 1.0],
+                "input": "/1",
+                "output": "coord_sys0"
               }
             ]
           },
@@ -63,7 +83,9 @@
             "coordinateTransformations": [
               {
                 "type": "scale",
-                "scale": [1.0, 1.0, 2.0, 2.0, 2.0]
+                "scale": [1.0, 1.0, 2.0, 2.0, 2.0],
+                "input": "/2",
+                "output": "coord_sys0"
               }
             ]
           }
@@ -71,10 +93,11 @@
         "coordinateTransformations": [
           {
             "type": "scale",
-            "scale": [0.1, 1.0, 1.0, 1.0, 1.0]
+            "scale": [0.1, 1.0, 1.0, 1.0, 1.0],
+            "input": "coord_sys0",
+            "output": "coord_sys1"
           }
         ],
-        "type": "gaussian",
         "metadata": {
           "description": "the fields in metadata depend on the downscaling implementation. Here, the parameters passed to the skimage function are given",
           "method": "skimage.transform.pyramid_gaussian",
