+++ tests/data/examples/v06/labels_image_example.json
@@ -4,12 +4,27 @@
     "multiscales": [
       {
         "name": "example",
-        "axes": [
-          { "name": "t", "type": "time", "unit": "millisecond" },
-          { "name": "c", "type": "channel" },
-          { "name": "z", "type": "space", "unit": "micrometer" },
-          { "name": "y", "type": "space", "unit": "micrometer" },
-          { "name": "x", "type": "space", "unit": "micrometer" }
+        "coordinateSystems": [
+          {
+            "name": "coord_sys0",
+            "axes": [
+              { "name": "t", "type": "time", "unit": "millisecond" },
+              { "name": "c", "type": "channel" },
+              { "name": "z", "type": "space", "unit": "micrometer" },
+              { "name": "y", "type": "space", "unit": "micrometer" },
+              { "name": "x", "type": "space", "unit": "micrometer" }
+            ]
+          },
+          {
+            "name": "coord_sys1",
+            "axes": [
+              { "name": "t", "type": "time", "unit": "millisecond" },
+              { "name": "c", "type": "channel" },
+              { "name": "z", "type": "space", "unit": "micrometer" },
+              { "name": "y", "type": "space", "unit": "micrometer" },
+              { "name": "x", "type": "space", "unit": "micrometer" }
+            ]
+          }
         ],
         "datasets": [
           {
@@ -17,7 +32,9 @@
             "coordinateTransformations": [
               {
                 "type": "scale",
-                "scale": [1.0, 1.0, 0.5, 0.5, 0.5]
+                "scale": [1.0, 1.0, 0.5, 0.5, 0.5],
+                "input": "/0",
+                "output": "coord_sys0"
               }
             ]
           }
@@ -25,7 +42,9 @@
         "coordinateTransformations": [
           {
             "type": "scale",
-            "scale": [0.1, 1.0, 1.0, 1.0, 1.0]
+            "scale": [0.1, 1.0, 1.0, 1.0, 1.0],
+            "input": "coord_sys0",
+            "output": "coord_sys1"
           }
         ]
       }
